parte2/requirements.txt
# Núcleo ETL
pandas
SQLAlchemy
PyMySQL
cryptography
# Config y CLI
python-dotenv
click

# Logs (opcional pero útil)
loguru
